refactor: auth middleware refactoring
